+++ src/components/Form/components/FormEditButton.component.tsx
@@ -4,26 +4,33 @@ import { useFormContext } from "../Form.context";
 import { formEditButtonStyles } from "../Form.styles";
 
 import { Button } from "@/components/Button";
+import { useStatic, useUikitTranslation } from "@utils";
 
 import type { FormEditButtonStyleProps } from "../Form.types";
 
 /**
  * Renders a form's edit button and hides it based on edit state.
  *
+ * @version 0.0.2
+ * @private
+ *
  * @return {JSX.Element} The rendered form edit button component.
  */
 export const FormEditButton = (): JSX.Element => {
   const { isEditing, startEditing, readonly } = useFormContext();
+  const tl = useUikitTranslation();
+
+  const label = useStatic(tl.action.edit);
 
   return (
     <EditButton
       $hidden={isEditing || readonly}
       onClick={startEditing}
       icon="Pen"
-      title="Edit"
+      title={label}
       testId="form-edit-button"
       overrideTestId
-      label={"Edit"}
+      label={label}
       secondary
       expandOnHover
     />